(WIP) Move root bits to standalone file
